Update rest of base person model classes
